Add support for customized name mapping
