+++ pymtl3/passes/sverilog/__init__.py
@@ -1,3 +1,4 @@
 
-from .import_.ImportPass import ImportConfigs, ImportPass
+from .import_.ImportConfigs import ImportConfigs
+from .import_.ImportPass import ImportPass
 from .translation.TranslationPass import TranslationPass
